+++ lib/src/features/auth/domain/repositories/auth_gateway.dart
@@ -1,0 +1,7 @@
+import '../entities/auth_status_snapshot.dart';
+
+abstract interface class AuthGateway {
+  Future<AuthStatusSnapshot> loadStatus();
+
+  Future<AuthStatusSnapshot> signOut();
+}